requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pyodbc==5.2.0
anthropic==0.25.0  # For Claude API
requests==2.32.3   # For HTTP requests to FastAPI